sc: store sort order and reverse the sort for default view
We store the sort order of the default view in SheetViewManager,
so we can reverse the sort when the sheet view is changed.

Tested scenarios are when the auto-filter is sorted in the default
view, and the sheet view is changed. Additional scenario is when
the auto-filter is sorted again (for example descending following
with the ascending order) and the changing the sheet view should
still work correctly.

Change-Id: Ic42e8a51d6ce17db6b2df80bafb62f6f8c48a6a8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190599
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins