tdf#89557 Set the movement accelerators to expected ones on MacOS
On MacOS, the keys to move to the beginning and end of the document are
expected to be ⌘↑ and ⌘↓. Similarly, to move between paragraphs, Alt+↑
and Alt+↓ are supposed to be used. This does work correctly in
LibreOffice via the doCommandBySelector method in salframeview.mm. The
MacOS SDK itself handles translating the key events into selectors such
as “moveToBeginningOfDocument”. These are then passed to
doCommandBySelector and LibreOffice has an implementation of each
interesting selector.

However, when the cursor is in a table of contents in a document, the
frame doesn’t have the InputContextFlags::Text option and the selector
translation code is bypassed. Instead the last known key event is
sent. For ⌘↑ and friends, by default this gets interpreted as an
accelerator to move between paragraphs like Ctrl+↑ on other
platforms. That means you can’t move to the beginning and end of the
document in the expected way on MacOS if the cursor is in a table of
contents.

This patch tries to fix that by just reassigning the default
accelerators to be what they are expected to be on MacOS. That way if
anything stops the translation to a selector from working then the usual
accelerator code will work correctly.

Sadly it still isn’t possible to reassign these accelerators to
something else because the selector code will still end up overriding
them when the cursor isn’t in a table of contents. doCommandBySelector
is actually deprecated since MacOS 10.6 so in the long run this will
probably have to be done a different way. In the short term we can’t
just remove the handlers for moveToBeginningOfDocument etc because it
breaks movement in text input fields in dialog boxes. Either way, this
patch improves the situation for MacOS users because at least if you
look at the default accelerator list in the customization dialog then
the keys will now actually do what it says they do.

These changes were made with the following bit of Emacs Lisp code:

(defun fix-accelerator (start end old-url new-url &optional add-key add-url)
  (let ((end (copy-marker end)))
    (goto-char start)
    (when (re-search-forward (concat "^            <value\\b[^^]*\\(>\\)"
                                     (regexp-quote old-url)
                                     "</value>")
                             end
                             t)
      (beginning-of-line)
      (let ((value-node (buffer-substring (point) (match-beginning 1)))
            (node-start (buffer-substring start (match-beginning 1))))
        (goto-char (match-beginning 1))
        (insert " install:module=\"unxwnt\"")
        (forward-line)
        (insert value-node
                " install:module=\"macosx\">"
                new-url
                "</value>\n")
        (when add-key
          (goto-char end)
          (forward-line)
          (insert "        <node oor:name=\""
                  add-key
                  "\" oor:op=\"replace\">\n"
                  node-start
                  " install:module=\"macosx\">"
                  add-url
                  "</value>\n"
                  "          </prop>\n"
                  "        </node>\n"))))
    (set-marker end nil)))

(defun fix-accelerators-region (start end)
  (let ((end (copy-marker end))
        (replacements '(("DOWN_MOD1"
                         ".uno:GoToNextPara" ".uno:GoToEndOfDoc"
                         "DOWN_MOD2" ".uno:GoToNextPara")
                        ("DOWN_SHIFT_MOD1"
                         ".uno:EndOfParaSel" ".uno:EndOfDocumentSel")
                        ("DOWN_SHIFT_MOD2"
                         ".uno:MoveDown" ".uno:EndOfParaSel")
                        ("UP_MOD1"
                         ".uno:GoToStartOfPara" ".uno:GoToStartOfDoc"
                         "UP_MOD2" ".uno:GoToStartOfPara")
                        ("UP_SHIFT_MOD1"
                         ".uno:StartOfParaSel" ".uno:StartOfDocumentSel")
                        ("UP_SHIFT_MOD2"
                         ".uno:MoveUp" ".uno:StartOfParaSel")))
        (case-fold-search nil))
    (while replacements
      (let ((key (caar replacements))
            (replacement-data (cdar replacements)))
        (goto-char start)
        (when (re-search-forward (concat "^        <node oor:name=\""
                                         (regexp-quote key)
                                         "\"")
                                 end
                                 t)
          (forward-line)
          (let ((start (point)))
            (re-search-forward "^        </node>")
            (beginning-of-line)
            (apply 'fix-accelerator start (point) replacement-data)))
        (setq replacements (cdr replacements))))
    (set-marker end nil)))

(defun fix-accelerators ()
  (interactive)
  (let ((groups '("text.TextDocument"
                  "text.GlobalDocument"
                  "text.WebDocument"
                  "xforms.XMLFormDocument"
                  "sdb.FormDesign"
                  "sdb.TextReportDesign"))
        (case-fold-search nil))

    (let ((n groups))
      (while n
        (goto-char (point-min))
        (re-search-forward (concat "^      <node oor:name=\"com.sun.star."
                                   (regexp-quote (car n))
                                   "\""))
        (forward-line)
        (let ((start (point)))
          (re-search-forward "^      <")
          (beginning-of-line)
          (fix-accelerators-region start (point)))
        (setq n (cdr n))))))

Change-Id: I81d7ebeaa324f1e96f70b58916f71d453a3e8f38
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190600
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>