sc: fix scrolling in sheet view, add check if a tab no. is valid
Scrolling needs to use the tab number of the actual sheet we are
currently using in the UI. The same goes for the grid.

Added a function to check if a tab number is currently valid, to
reduce duplication and to make the code a bit more readable.

Change-Id: I4e76a61efaa3632b020cfd75ea77c0809f08f6da
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190598
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
(cherry picked from commit 202a1fead00d235c2b7e330d906e852a0e597380)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190945
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>