tdf#100287 svx dialog: sync "first header" with "first footer"
When either the header OR the footer wants different first content,
then (by design) the application applies it to both of them.
(True for Writer only - not true for Calc!!!)

It is also important (for the user experience) that either one
is able to stop using first content altogether (for Writer only).

This was already working to some extent,
but the user experience was not consistent or predictable.

Specifically, the improvements with this patch:
-provide an appropriate "default" value when turning on a H/F.
-always visually match the most recent "first content" choice.
-in Calc, simply visiting the H/F no longer turns on first page.

The prior situation was very "fickle",
and "first content" could be affected
without the user even touching that box (for both Writer and Calc).

The underlying reason this patch can work is because of
SvxHFPage::SvxHFPage SetExchangeSupport().
That causes SfxTabDialogController::DeactivatePage
to update the m_xExampleSet which is then passed
to SvxHFPage::ActivatePage.
Thus the active tab knows about all the dialog's current settings.

Since Reset() calls ActivatePage,
it was safe to move the initialization out of Reset.

make -srj1 UITest_calc_tests8 \
    UITEST_TEST_NAME=tdf144940.tdf144940.test_tdf144940 \
    SAL_USE_VCLPLUGIN=gen

Change-Id: Icd6084ca50d244317c807266f04c85dadbba2bb1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190606
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
(cherry picked from commit 11cd6b2fbc86e28506d5b26a436df1b4b2b5bf09)
Reviewed-on: https://gerrit.collaboraoffice.com/c/core/+/4419
Tested-by: Jenkins CPCI <releng@collaboraoffice.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>