tdf#168175 Use sort parameter language in natural sort
Now the natural sort uses the language given in the sort parameter. Only
if it does not contain a language, then the global locale is used. The
language determines, which character is used as decimal separator.

This corresponds to ODF where both the locale info and whether to use
natural sort belong to the database range attributes.

Change-Id: I52629bdff7f6a93974770d2922e80ac95df21a5d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190622
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>