sw markdown export: handle block quote
<https://spec.commonmark.org/0.31.2/#example-228> says the quote block
markup should be written first, then the heading, so do it in that
order.

Also fix the import side: set the paragraph style in
SwMarkdownParser::AddBlockQuote(), similar to how
HtmlTokenId::BLOCKQUOTE_ON is handled in SwHTMLParser::NextToken(), so
the export side has an easier time with recognizing this is a block
quote.

Change-Id: I305c7a31832cba44cb66943a5fc45a966ad62d55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190654
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>