tdf#168317 sw markdown export: handle code block
The bugdoc has a multi-paragraph code block, import that to Writer, save
as markdown, the text is written as normal text.

Note that unlike block quote or heading, there is no marker at the start
of all paragraphs: only before the first and after the last one.
However, in the Writer doc model, this is a per-paragraph setting for
the used style.

Fix the problem by checking if the para style is RES_POOLCOLL_HTML_PRE,
then seeing if there is a similar paragraph before/after, so we only emit
the start/end marker only once, even for multi-paragraph code blocks.

Note that our own markdown import currently creates single-paragraph
code blocks by replacing paragraph breaks with newlines.

Change-Id: I0803bc71c758f250489d11e90b7f090966c8a8cc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190687
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>