Resolves: tdf#152129 Save/IsChanged for the SvxRectCTL
looks to me that we should make this behave like other controls
and generally set the m_eDefRP when explicitly asked to save
so that "Reset" resets back to that saved value.

And then we can check if the control is "changed" by comparison
of m_eDefRP vs m_eRP and use that to apply the properties.

Change-Id: Icef87fb7e3dd653267e6882f4e8d22adb3e0ff68
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190694
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>