tdf#168278 editeng a11y: Normalize selection start/end index
In AccessibleEditableTextPara::GetSelection, call
ESelection::Adjust so that the earlier position in
the text is considered the start position and the
position later in the text is considered the end
position, independent of whether the selection was
done forwards or backwards.

Always using the smaller index for the start index
and the greater index for the end index matches what
is done for Writer and what is also seen for other
applications as well.

Example results for typing "abcd", moving the mouse to
after the "c" and selecting two characters backwards
(i.e. "bc" is selected, with the caret being in
front of the "b":

Writer, with either the gtk3 or qt6 VCL plugin
and the paragraph selected in Accerciser's tree view
of the LO a11y hierarchy gives this in the IPython
console:

    In [2]: acc.queryText().getSelection(0)
    Out[2]: (1, 3)

For a Calc cell in edit mode, the result would
previously be:

    In [4]: acc.queryText().getSelection(0)
    Out[4]: (3, 1)

, but now it matches the behavior seen with
Writer.

In a quick test with Gedit on Linux and the
URL bar on Firefox on Windows and querying
the selection using the IAccessibleText
selection, the lower index was always reported
for the start index as well and the greater one
for the end index as well.

For gtk3, this additionally ensures that a selection
count of 1 is reported for a non-zero selection
that was done backwards as well, as
text_wrapper_get_n_selections returns 0 if the
end index isn't larger than the start index.

Change-Id: I549d85968f62cbfc2ee4d39c5a4aeba8767df14d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190716
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>