tdf#152519 qt: Update more IM query atttributes
So far, QtWidget::handleKeyEvent was only
notifying that the cursor rectangle (Qt::ImCursorRectangle)
has changed.
But more attributes can change, which were previously
not reported. In particular the cursor position
(Qt::ImCursorPosition) is relevant for the tdf#152519
scenario of fcitx5-unikey. (Many thanks to Xuetian Weng for
the insights in tdf#152519 comment 9!)

As the QInputMethod::update doc [1] says:

> In particular calling update whenever the cursor position changes is
> important as that often causes other query attributes like surrounding
> text and text selection to change as well. The attributes that often
> change together with cursor position have been grouped in
> Qt::ImQueryInput value for convenience.

Use this Qt::ImQueryInput, which is one step
towards making the tdf#152519 scenario work.

However, as mentioned in tdf#152519 comment 9, this
isn't sufficient, since not all values reported by
QtWidget::inputMethodQuery are correct.
This will be addressed separately.

[1] https://doc.qt.io/qt-6/qinputmethod.html#update

Change-Id: Ic707cd6c7e239798d034ec0e2cb11499b60f89e9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190742
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins