sw markdown export: handle table cell adjustment
The bugdoc has a table where all lines have the same amount of cells,
and the "columns" are left, center and right aligned. All alignment is
set to "left" on markdown export.

The problem is that Writer has no way to represent cell-level horizontal
alignment, but the HTML parser sets the cell alignment on the content of
the cell.

Fix the problem by doing the same on markdown export: take the first
text node in the cell, and if that has a custom alignment, then work
with that.

Interestingly this cell-level horizontal alignment is not only missing
from ODF, it's also missing from OOXML.

Change-Id: Ibf0cfd1e947f342e8745e6198742882893f08fe3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190739
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>