tdf#152519 vcl: Report IM cursor position
Based on previous commit

    Change-Id: I46fdff87e84fbbe62f9b60b50376b6123d59fdac
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Sep 10 10:53:18 2025 +0200

        tdf#152519 Distinguish cursor/anchor for IM selection

, add SalSurroundingTextRequestEvent::mnCursorPos to report
the cursor position as well in such an IM request.

Instead of normalizing the Selection, use Selection::End()
for the cursor position (see above-mentioned commit for more
details) and std::min/std::max to identify the start/end
index of the selection. Semantic for start/end index of the
remains the same, so this commit shouldn't result in any change
of behavior yet.

This commit prepares for porting
QtWidget::inputMethodQuery/lcl_retrieveSurrounding
from using a11y API to using SalEvent::SurroundingTextRequest,
similar to what

    commit ce5e41ab99af350ca8f4b9fef3017d53f3526f83
    Author: Caolán McNamara <caolanm@redhat.com>
    Date:   Sun Oct 25 15:14:56 2020 +0000

        Related: tdf#137620 use existing SalEvent::SurroundingTextRequest

did for gtk3.

Change-Id: I600b834737e4e46a60491b91e64f8f8bc4ad806c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190747
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins