tdf#160838 tdf#130857 qt weld: Signal on popup close
Call weld::Popover::signal_closed to notify
when the popup is closed.

To do so, handle the QEvent::Close event
in QtInstancePopover and queue the signalling
of the closing for when control returns to
the main loop.

At the point in time that QtInstancePopover::eventFilter
gets called with the event, the popup hasn't
actually been closed yet. This only happens
when the event gets processed by the widget
later, so calling Popover::signal_closed()
right away would be too early.

Sending the signal is e.g. needed to allow opening the
comment popup in Impress multiple times
in a WIP branch where support for native Qt widgets
for that popup is declared with
SAL_VCL_QT_USE_WELDED_WIDGETS=1.

Otherwise, AnnotationPopup::closePopup would
not get called, `AnnotationPopup::mpAnnotationWindow`
wouldn't be reset and therefore not newly set
and shown when clicking on a comment in Impress again.

Relevant part of the backtrace when the signal is sent
as it should:

    1   sd::AnnotationPopup::closePopup
    2   sd::AnnotationPopup::PopupModeEndHdl
    3   sd::AnnotationPopup::LinkStubPopupModeEndHdl
    4   Link<weld::Popover&, void>::Call
    5   weld::Popover::signal_closed
    [...]

(More is needed before support for that popup
can actually be declared in
QtInstanceBuilder::IsUIFileSupported; this is
just one step.)

Change-Id: I472c101f7871a5189b985a5d89e642287450ec5d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190775
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>