sw: PDF export: fix crash on section in table in section
soffice.bin: vcl/source/gdi/pdfextoutdevdata.cxx:798: void vcl::PDFExtOutDevData::EndStructureElement(): Assertion `mpGlobalSyncData->mCurrentStructElement != 0' failed.

The problem is that the parent-opening code in
SwTaggedPDFHelper::BeginBlockStructureElements() when called for a
section in a table cell opens a section that is outside the table.

In this case, there are actually 2 section frames, one containing the
table and one in the table cell, so reopening the outer section frame
that is already on the stack ends up restoring a 0 parent SE and that
triggers the assert.

(regression from commit d5f68529a79c615f989fcfeef248d887a6e10f5a)

Change-Id: Idd5be655345f96e1c2de84c5623f6e1dba4aae1b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190781
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins