tdf#168263 editeng: Fix wrong max ascent for text without ruby
Editeng must adjust the maximum ascent of lines to accommodate ruby
characters. This adjustment is only necessary when ruby characters are
actually used, and the intention was to only apply the adjustment in
that case. Unfortunately, due to an implementation error, this
adjustment was instead performed for any line containing formatting.

Change-Id: Ibd4ea3cda8b40c95f41f1d84b47155455a20a1cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190784
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins