Use a bit more standard null date
The two places in cellvalueconversion.cxx using the null date were
introduced in commit a7e949673cda3f0e2abfa49b2c22735f33205d01
(gridfixes: #i117265# use a NumberFormatter for converting cell
values to text., 2012-09-14); they are used together, so that the
code in lcl_convertDateToDays uses the same epoch as defined in
the respective number formatter, and the obtained number value is
only ever used inside convertToString in a call to the formatter's
convertNumberToString. The old value was arbitrary; and the new
value has some optimizations in the code (see Date::DateToDays and
comphelper::date::convertDateToDaysNormalizing called from it).

Change-Id: I9c050d26ace1cf3fccde579d0d91c00f26bb3fd8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190788
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>