tdf#168341 sw markdown filter: handle links on images
The bugdoc has an image which has a link: the link is lost on both
import and export.

The doc model for links is different for text an images: the text is
covered by an SwFormatINetFormat hint, but images have their
SwFlyFrameFormat, and that can contain an SwFormatURL.

Fix the problem by looking at the currently pending attributes stack
while inserting images: if a link is open, also set that on the image.
The original link will be discarded, because it covers no text.

Also fix the export to write markup when the fly format has an
SwFormatURL.

Change-Id: Ib94bcd0903835d6eb197ef2077075390305f37c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190787
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>