sc: don't remove the sheet view entry from the vector
SheetViewID is directly related to the index in the vector where
we store the sheet views, so it could be catastrphic if we remove
the entry from the vector as SheetViewID would not reference the
correct SheetView. Instead just reset the shared_ptr, so it has
not SheetView, but this will create a hole in the vector. It's
however unlikely that this is a problem as we will not have a lot
of sheet views for a table.

Also remove duplication when checking if the SheetViewID is valid
(inside the bounds of the vector).

Change-Id: Ifd4c67f49def942d9b9afdd2fc63f3a690124457
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190889
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>