tdf#46623 sw PageBreakWin: always move cursor to FindFirstBodyContent
When activating the blue-line page break
decorator menu (edit or delete),
the page break would be disabled if the cursor
was not in FrameTypeFlags::BODY (or in a table).

Well, the cursor position doesn't matter
when the mouse activates this SwBreakDashedLine,
so just always do what the table had to do anyway
and that is to move the cursor
so that a call to rSh.GetFrameType can return BODY.

This allows SwBreakDashedLine to work
when the cursor is in the header/footer/footnotes/flies etc.

Change-Id: I64b49bed13821b46c3db2b132edeeb798dd2c3ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190828
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>