tdf#130857 tdf#160838 qt weld: Support inserting check menu items
Looking into the other weld::Menu implementations shows
that TRISTATE_TRUE for param eCheckRadioFalse means that
a check menu item should be inserted.
Implement handling for that.

This will e.g. be used by the menu in the comment
popup in Impress.

To trigger the menu:

* start Impress
* "Insert Comment"
* right-click on the comment area

This comment is one step towards declaring support
for using native widgets there, but more work is
needed.

Change-Id: Ic4f594dbe3c2b5bcbaaa1d458fa6ae610ec24ba4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190833
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins