tdf#130857 tdf#160838 qt weld: Convert accelerator in menu item text
This will e.g. be used by the menu in the comment
popup in Impress.

To trigger the menu:

* start Impress
* "Insert Comment"
* right-click on the comment area

This doesn't actually use native Qt widgets yet,
but this commit is one step towards declaring support
for using native widgets there. More work is
still needed before that can be done.

In a WIP branch without this commit in place,
menu entries with literal text "Cop~y" and "~Paste"
were seen in the menu.
Now they are displayed as expected.

Change-Id: Ic7071b3aa41c98e121f7b325e70070cc31dd6129
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190835
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>