tdf#130857 tdf#160838 qt weld: Conditionally support sd comment menu
Declare conditional support for the menu shown for
the Impress comment popup, i.e. declare it as supported
when the parent is a native Qt widget as well.

While the comment popup is currently not supported
with Qt widgets yet, this commit prepares for the
point in time when it is, at which point the
menu will also be a native Qt menu.

The menu already works as expected in a WIP branch
experimenting with using native widgets for that
popup with SAL_VCL_QT_USE_WELDED_WIDGETS=1, but
more work is needed for the popup itself before
that one can also be declared as supported for real.

To trigger the menu:

* start Impress
* "Insert Comment"
* right-click on the comment area (or alternatively,
  type and select some text and press the context menu
  button)

Change-Id: Ie8a1b045c2efe39d645b96dec6ddcabef399da45
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190836
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins