tdf#168321 sw format redline, direct: fix ODF strict export
Save the bugdoc in strict ODF, the exporter crashes due to an assertion
failure.

What happens is that strict ODF doesn't add the loext namespace alias to
the alias list, so we crash when we would write an loext attribute.

Fix the problem by simply throwing away the direct format of a format
redline when saving to strict ODF, just to please the validator.

If we're at it, also don't register the direct format of the format
redline in the autostyle pool in that case, since it would be
unreferenced anyway.

Change-Id: I0de9ee114be9ef67f97716cf5fd9aaa9f8757683
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190844
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>