Add hyperlink dialog document tab to online
We hide a lot more than on other tabs. The main idea is to allow
bookmarks to be linked to, but this does add other behaviour.
For example, linking to a shape using 'Shape 1|drawingobject' to
link to a shape. When I tested this in online, this behaviour was
very inconsistant and slow, although this could be related to
document size.

Change-Id: I6a6a696443a677cc54aa4a62594b4b6c6aa4d114
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190895
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>