sc: unsort when synching a change from sheet view to default view
If we sort in sheet view and change the value in sheet view, we
need to reverse the sort when syncing the value with the default
view. When a sheet view is sorted, we store the order inside
SheetView. If we sort again, we need to combine the existing sort
order with the new sort order (sort asc. and then sort desc).

We also need sheet view info in ScTable of a sheet view, so we can
properly set the sort order. This sent the information when we
create the sheet view.

Change-Id: I4d7f2c8add3f137025bcc9c6f29482795fa32ccb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190891
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/197819
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>