vcl: rename DrawBitmapEx() to DrawAlphaBitmap()
The DrawBitmapEx() functions forward to this version of DrawBitmapEx().
Before they foward they check if it is an alpha bitmap, if it isn't then
they call on DrawBitmap().

The version of the function they forward to merely draws an alpha
bitmap - there is no need for it to check that the bitmap has an alpha
channel as this was done already. In fact, anything that calls directly
on this function without such a bitmap should be calling DrawBitmapEx(),
so I've added an assert() to flag when this occurs.

Change-Id: Ia94caf28ba1d49ec7c9f17654ca73ef51baf49a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190896
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>