sc: introduce SortOrderReverser to handle sort order conversion
SortOrderReverser class handles the reversing of the sort order,
which is needed on the sheet view and default view level and is
more or less duplicated code in SheetView and SheetViewManager.
With SortOrderReverser it is possible to remove the duplication
and write a unit test just to test reversing the sort.

Change-Id: I7a0726bf71e0b25bcf68403b7c13c3dfac889dd5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190910
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/200014
Tested-by: Tomaž Vajngerl <quikee@gmail.com>