sc: synching when default view and sheet view are sorted
In this scenario we have sheet view and the default view sorted
in different ways, so we need to first unsort the sheet view sort,
and then resort the default view sort so we get the correct cell
address of the cell that we are changing.

Resort is the reverse of unsort. We need to search where in the
sort order array is the value of the input row. The array index
will be the result and the first row + index will give us the
resorted row address.

Change-Id: I08dd7f6fc5593464cb2e6bd36ffb3ff250cebf8d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190912
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>