Handle removing shortcuts in gtksalmenu
When a menu item has no shortcut its accelerator key code is set to
zero. When the GTK backend encounters this it was just ignoring the key
code and doing nothing to update the menu item. This meant that if an
accelerator is changed from one menu item to another then removing the
shortcut from the old menu item wouldn’t work. This patch fixes it to
handle that case.

The line calling g_strcmp0 looked suspicious. It was only called if
aCurrentAccel was nullptr, in which case it’s not worth calling because
a check whether aAccelerator is not nullptr would have the same effect.
I wonder if it was meant to be || instead of &&. In either case
g_strcmp0 is meant to handle null pointers so it seems like it should do
the right thing without having to check if aCurrentAccel is null.

This fixes the issues with the GTK backend mentioned in comment #8 of
tdf#144560.

Change-Id: Ica63a8666d2a4ca79a0c566dd43e7cb53ec9df67
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190930
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins