tdf#144560 Handle removing shortcuts in osx/salmenu
When a menu item has no shortcut its accelerator key code is set to
zero. When the MacOS backend encounters this it was just ignoring the
key code and doing nothing to update the menu item. This meant that if
an accelerator is changed from one menu item to another then removing
the shortcut from the old menu item wouldn’t work. This patch fixes it
to handle that case.

Change-Id: I38fa09958b78b9cbc4ea8fb9b222d4671c91fb92
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190929
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>