ScriptForge Extend Python-Basic gateway for objects
A Python ScriptForge object may be passed to
Basic functions, so far only when it is
positioned as the first argument in the
method signature.

Like in:
   dialogB.Center(dialogA)

The actual commit extends the protocol
between Python and Basic to allow passing
Python objects from another position
in the method signature.

However this possibility is limited to
maximum 1 such argument.
(Note that removing this limitation,
if ever needed, would not require
a big change.)

No impact on user existing scripts.
No documentation update.

Change-Id: I191b7bb7c32b158b7274815577349ddfa1406b93
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190931
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>