sw markdown export: handle line breaks
Export the bugdoc as markdown, the line break is lost when opening the
result with a markdown viewer.

<https://spec.commonmark.org/0.31.2/#hard-line-breaks> says the markup
to be used is a normal line break that "is preceded by two or more
spaces", which explains why just writing the newline character as-is
wasn't enough.

Fix the problem in OutEscapedChars() where we can emit the "escaped"
line break as a special case.

The alternative markup would be a backslash at the end of the line,
which is more readable, but seems that's less supported by parsers (e.g.
okular and https://markdowntohtml.com/ doesn't support it).

Change-Id: Iedabe7ef2b42027f90595bf76e952253141d9b3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190940
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>