remove git from VS modules (use winget copy instead matching wiki docs)
the versions differ now, and when you try to apply the configuration
from within git-bash the Visual Studio installation fails because it
wants to replace bash that is still-in-use.

Also fix a bug in the cloud-init config for the wsl container,
locked_passwd defaults to true in 24.04, and that effectively disables
the password, rendering sudo inoperable. (could still run
"wsl.exe -u root" to login as root and do admin tasks that way)

Alternatively the user could be reconfigured to allow for passwordless
sudo, but that's a matter of preference, I'm just personally so used to
always use sudo with password/it is a nice reminder to only change that
if you know what you're doing/that the container was installed for a
specific purpose and the user might want to install a second copy for
other stuff.

PS:
On Windows Server 2025 the installation using the ubuntu2404.exe alias
won't work/has to be done manually using
    wsl.exe --install --distribution Ubuntu-24.04
need some further testing whether that's wider change or specific to
Windows Server

Change-Id: Ie73fdd14cc692aa22e6bc1de5e7465ca60f8a0cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190970
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins