ScriptForge (Sharedmemory) review StoreValue() signature
Thanks to commit
  https://gerrit.libreoffice.org/c/core/+/190931
  (Extend Python-Basic gateway for objects)
a signature with an object instance as argument
is allowed in any position, i.o. the 1st only.

This is applicable to
  sharedmemory.StoreValue(value, variablename)
Indeed, 'value' may be such an object instance.

The signature becomes now
  sharedmemory.StoreValue(variablename, value)
which is more intuitive.

No impact on user scripts <= 25.8

Documentation should take both commits
into account.

Change-Id: I1a07cd740fc9dcba1077932bba0f69735a7bc647
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190979
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>